+++ src/components/Controls/IntervalSlider/IntervalBorders.tsx
@@ -1,3 +1,4 @@
+import { cn } from '@helpers/tailwind';
 import { Rect } from '@types';
 import { INTERVAL_BORDER_WIDTH } from './constants';
 
@@ -5,19 +6,23 @@ interface IntervalBordersProps {
   intervalStartX: number;
   intervalEndX: number;
   trackArea: Rect;
+  isDisabled?: boolean;
 }
 
 export const IntervalBorders = ({
   intervalStartX,
   intervalEndX,
-  trackArea
+  trackArea,
+  isDisabled = false
 }: IntervalBordersProps) => {
   return (
     <>
       <div
-        className='absolute'
+        className={cn('absolute', {
+          'bg-c7': !isDisabled,
+          'bg-c1': isDisabled
+        })}
         style={{
-          backgroundColor: 'var(--c7)',
           top: 0,
           left: intervalStartX,
           width: intervalEndX - intervalStartX,
@@ -25,9 +30,11 @@ export const IntervalBorders = ({
         }}
       />
       <div
-        className='absolute'
+        className={cn('absolute', {
+          'bg-c7': !isDisabled,
+          'bg-c1': isDisabled
+        })}
         style={{
-          backgroundColor: 'var(--c7)',
           top: trackArea.y + trackArea.height,
           left: intervalStartX,
           width: intervalEndX - intervalStartX,