+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 
+### Fixed
+- interval editor disabled when pad is empty
+
+
 ## v1.4.1 (2025-03-05)
 
 ### Changed